pyproject.toml
[project]
name = "file-server"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "argon2-cffi>=25.1.0",
    "fastapi[all]>=0.135.1",
    "passlib[bcrypt]>=1.7.4",
    "python-dotenv>=1.2.2",
    "python-jose[cryptography]>=3.5.0",
]
